+++ Blinkiis_Portraits/portraits/party.lua
@@ -65,7 +65,6 @@ function BLINKIISPORTRAITS:InitializePartyPortrait()
 				portraits[unit].lastGUID = nil
 				portraits[unit].forceUpdate = true
 
-				BLINKIISPORTRAITS:UpdateSettings(portraits[unit], BLINKIISPORTRAITS.db.profile[type])
 				BLINKIISPORTRAITS:UpdateTexturesFiles(portraits[unit], BLINKIISPORTRAITS.db.profile[type])
 				BLINKIISPORTRAITS:UpdateSize(portraits[unit])
 				BLINKIISPORTRAITS:UpdateCastSettings(portraits[unit])