+++ Blinkiis_Portraits/portraits/pet.lua
@@ -66,7 +66,6 @@ function BLINKIISPORTRAITS:InitializePetPortrait()
 			portraits[unit].lastGUID = nil
 			portraits[unit].forceUpdate = true
 
-			BLINKIISPORTRAITS:UpdateSettings(portraits[unit], BLINKIISPORTRAITS.db.profile[type])
 			BLINKIISPORTRAITS:UpdateTexturesFiles(portraits[unit], BLINKIISPORTRAITS.db.profile[type])
 			BLINKIISPORTRAITS:UpdateSize(portraits[unit])
 			BLINKIISPORTRAITS:UpdateCastSettings(portraits[unit])